# openapi.yaml
openapi: 3.0.0
info:
  title: saludconecta API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3015/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /paciente:
    get:
      tags:
        - Paciente
      summary: Listar todos los Pacientes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Pacientes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Paciente
      summary: Crear Paciente
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Paciente creado
  /paciente/{id}:
    get:
      tags:
        - Paciente
      summary: Obtener Paciente por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Paciente encontrado
        '404':
          description: Paciente no encontrado
    patch:
      tags:
        - Paciente
      summary: Actualizar Paciente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Paciente actualizado
    delete:
      tags:
        - Paciente
      summary: Eliminar Paciente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Paciente eliminado
  /paciente/{id}/upload-image:
    post:
      tags:
        - Paciente
      summary: Subir imagen a Paciente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
  /medico:
    get:
      tags:
        - Medico
      summary: Listar todos los Medicos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Medicos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Medico
      summary: Crear Medico
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Medico creado
  /medico/{id}:
    get:
      tags:
        - Medico
      summary: Obtener Medico por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Medico encontrado
        '404':
          description: Medico no encontrado
    patch:
      tags:
        - Medico
      summary: Actualizar Medico
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Medico actualizado
    delete:
      tags:
        - Medico
      summary: Eliminar Medico
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Medico eliminado
  /medico/{id}/upload-image:
    post:
      tags:
        - Medico
      summary: Subir imagen a Medico
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...
  /tratamiento:
    get:
      tags:
        - Tratamiento
      summary: Listar todos los Tratamientos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Tratamientos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Tratamiento
      summary: Crear Tratamiento
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Tratamiento creado
  /tratamiento/{id}:
    get:
      tags:
        - Tratamiento
      summary: Obtener Tratamiento por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tratamiento encontrado
        '404':
          description: Tratamiento no encontrado
    patch:
      tags:
        - Tratamiento
      summary: Actualizar Tratamiento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Tratamiento actualizado
    delete:
      tags:
        - Tratamiento
      summary: Eliminar Tratamiento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tratamiento eliminado
  /tratamiento/{id}/upload-image:
    post:
      tags:
        - Tratamiento
      summary: Subir imagen a Tratamiento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida